partition_manager: opt-in to use pm in single image build
Enable the configuration for zigbee which depends on PM.

Prior to this a single image build with a subsys that defined
a partition manager configuration would result in the
partition manager being invoked.

Instead, make the user opt-in to use partition manager in this
situation.

This is done to avoid breaking backwards compatibility and
require actions by downstream users.

Ref: NCSDK-5560
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>